Update nltk tokenizer block to also do pos tagging and lemmatizing
This was done, because doing everything in the same block is much easier
(pos tagging and lemmatizing in nltk requires lists sentences (where a
sentence is a list of words), and getting them out of the dataframe
again turned out to be a hassle. Moreover, for most other
tokenizers/lemmatizers the process of tokenizing/pos tagging/lemmatizing
is done in a single step anyway.